LayoutTests/fast/js/kde/const-expected.txt

PASS c is 11
PASS c is 11
PASS v is 1
PASS successfullyParsed is true
